refactor(dashboard): move the dashboard helpers and invite copy into the module — and DON'T rewrite the routine-prompt mirror (KAN-415 tail)
UI-Change-Approved: KAN-415

    src/lib/dashboard/{dismissal,profile-completion,resolve-widgets}.ts
                                    -> src/modules/dashboard/
    src/lib/invite-text.ts          -> src/modules/dashboard/invite-text.ts

Four byte-identical renames (100%). invite-text.ts is founder-owned copy; it
moves unchanged and is FOUNDER-OWNED at its new path, verified by running the
shipped is_protected there. CTL-035 flagged the old pattern dead before the fix,
as it did for beta-access/email.ts in the previous commit.

THE FINDING: A BLANKET PATH REWRITE DISARMS THE OUT-OF-REPO ATTESTATION

check-extraction-dod.sh carries ROUTINE_COUPLED — the paths named VERBATIM
inside claude.ai routine prompts. The check compares MOVED-FROM paths against
it, and its own error text explains the stake: "Luisa updates the prompt; no PR
can."

So the list is a mirror of text that lives OUTSIDE every repo CI can read. Its
entries are correct when they match what the PROMPT says — which, after a move
and before the prompt is edited, is the OLD path.

Rewriting an entry to the new path in the same commit that moves the file makes
the moved path stop matching, the gate report "not applicable", and the
attestation vanish. A commit that carries itself through the gate it is
disabling — structurally identical to the D9 finding, and it caught me twice:
the blanket rewrite hit this list in THIS commit and, unnoticed, in the previous
one for beta-access/email.ts.

Nothing detected it. check-extraction-dod.sh is not a CTL-035-registered
artefact, so a dead entry there is invisible; it was found by reading the diff.

Both entries are restored to the prompts' spellings, with a comment at the list
saying why they must not be rewritten and when they may be. The gate now fires
correctly for this PR, and the PR body carries the trigger ID it demands.

Note the in-repo guard is the opposite case and DOES move: is_protected in
check-ui-copy-ownership.sh describes the tree, so it points at
src/modules/dashboard/invite-text.ts. Same string, two artefacts, opposite
correct answers — one mirrors the repo, the other mirrors a prompt.

ESTATE

  modules.json               collapsed to the module root; the sed produced the
                             same nested-path duplication as the access commit
                             and the manifest's integrity test caught it again
  env-access-baseline.json   invite-text.ts rekeyed in place (one line)
  check-ui-copy-ownership.sh is_protected -> the new path
  source-paths manifest      regenerated; no key lost (checked by diffing the
                             key set before and after, not assumed — gotcha #31
                             is the case where one silently vanishes)
  archival records           the four dated research artefacts added to
                             ARCHIVE_FILES last commit were correctly skipped by
                             the sweep this time

VERIFICATION

tsc clean; CTL-035 0 dead / 155 live; is_protected FOUNDER-OWNED at the new
path; no stale reference to either old path outside the archival set.

Refs: KAN-415, KAN-428, KAN-411
