+++ docs/modularisation/KAN-422-dead-exports.md
@@ -101,7 +101,7 @@ cleanup items (§4).
 | # | File | Symbols | LOC | Tests to delete with it | Evidence |
 |---|---|---|---|---|---|
 | D1 | `src/app/_marketing/sections.tsx` | `Hero`, `ProfilePreview`, `HowItWorks`, `Sections`, `UseCases`, `WhatLyraIsNot`, `AboutLyra`, `ParentTeacherCallout`, `WishKnowFindFirstHand`, `CTA` | **376** | *none* | Only `AboutTrio` is imported (`src/app/(legal)/about/page.tsx:4`). The homepage inlines its own markup. Every other name's remaining repo hits are comments or unrelated identifiers. **Keep the file, keep `AboutTrio`.** |
-| D2 | `src/app/dashboard/share-profile.tsx` | `default` | **81** | *none* | KAN-154-B "Share your invite" client card. No importer; the only repo reference is a comment in `src/lib/invite-text.ts:7`. |
+| D2 | `src/app/dashboard/share-profile.tsx` | `default` | **81** | *none* | KAN-154-B "Share your invite" client card. No importer; the only repo reference is a comment in `src/modules/dashboard/invite-text.ts:7`. |
 | D3 | `src/lib/convene/google/calendar.ts` | `getFreeBusy` | **29** | *none* | Superseded by `src/lib/convene/calendar/google.ts::googleCalendarAdapter.getFreeBusy`, which is the one `adapterFor('google')` returns and `organise/actions.ts:188` calls. |
 | D4 | `src/lib/convene/google/oauth.ts` | `exchangeCodeForTokens` | 23 | `tests/unit/convene/google-oauth.test.ts` (**partial — see note**) | The Google callback route reimplemented this inline as `exchangeCodeForTokensVerbose` (`src/app/api/convene/oauth/google/callback/route.ts:61`). Sibling exports `buildAuthorizeUrl` + `refreshAccessToken` in the same file **are live** — so this test file covers live code and **cannot be deleted**, only the `exchangeCodeForTokens` describe-block. |
 | D5 | `src/lib/convene/invites/repository.ts` | `markInviteSent`, `markInviteFailed` | 21 | `tests/unit/…/invites-actions.test.ts`, `rsvp-submit-validation.test.ts` (**partial**) | No caller in any of the three repos; the dispatcher writes the same columns inline. Both test files also cover live code. |
