+++ src/modules/dashboard/resolve-widgets.ts
