+++ tests/support/source-paths.json
@@ -159,8 +159,6 @@
     "oauth": "src/lib/convene/microsoft/oauth.ts",
     "oauthConnections": "src/lib/convene/oauth-connections.ts",
     "postEvent": "src/lib/convene/post-event.ts",
-    "resolveWidgets": "src/lib/dashboard/resolve-widgets.ts",
-    "inviteText": "src/lib/invite-text.ts",
     "moderationPolicy": "src/lib/moderation-policy.ts",
     "pipeline": "src/lib/recommender/v2/pipeline.ts",
     "flags": "src/lib/retention/flags.ts",
@@ -173,6 +171,8 @@
     "providerGate": "src/modules/age/provider-gate.ts",
     "selfDeclaration": "src/modules/age/self-declaration.ts",
     "postLoginRedirect": "src/modules/auth/post-login-redirect.ts",
+    "inviteText": "src/modules/dashboard/invite-text.ts",
+    "resolveWidgets": "src/modules/dashboard/resolve-widgets.ts",
     "entitlementsService": "src/modules/features/entitlements-service.ts",
     "fileMagicBytes": "src/modules/guards/file-magic-bytes.ts",
     "sanitise": "src/modules/guards/sanitise.ts",
