+++ tests/support/source-paths.ts
@@ -172,8 +172,6 @@ export const SRC = {
   oauth: 'src/lib/convene/microsoft/oauth.ts',
   oauthConnections: 'src/lib/convene/oauth-connections.ts',
   postEvent: 'src/lib/convene/post-event.ts',
-  resolveWidgets: 'src/lib/dashboard/resolve-widgets.ts',
-  inviteText: 'src/lib/invite-text.ts',
   moderationPolicy: 'src/lib/moderation-policy.ts',
   pipeline: 'src/lib/recommender/v2/pipeline.ts',
   flags: 'src/lib/retention/flags.ts',
@@ -186,6 +184,8 @@ export const SRC = {
   providerGate: 'src/modules/age/provider-gate.ts',
   selfDeclaration: 'src/modules/age/self-declaration.ts',
   postLoginRedirect: 'src/modules/auth/post-login-redirect.ts',
+  inviteText: 'src/modules/dashboard/invite-text.ts',
+  resolveWidgets: 'src/modules/dashboard/resolve-widgets.ts',
   entitlementsService: 'src/modules/features/entitlements-service.ts',
   fileMagicBytes: 'src/modules/guards/file-magic-bytes.ts',
   sanitise: 'src/modules/guards/sanitise.ts',
